# server/docs/swagger.yaml
definitions:
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.8.9
paths:
...
  /limitedActivity/order/findOrder:
    get:
      parameters:
      - description: 订单ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 活动订单详情
      tags:
      - LimitedActivityOrder
  /limitedActivity/order/getOrderByCodePublic:
    get:
      parameters:
      - description: 订单号
        in: query
        name: code
        required: true
        type: string
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
                msg:
                  type: string
              type: object
      summary: 根据订单号查询活动订单(公开)
      tags:
      - LimitedActivityOrder
  /limitedActivity/order/getOrderList:
    get:
      parameters:
      - in: query
        name: activityId
        type: integer
      - in: query
        name: contactPhone
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - in: query
        name: orderNo
        type: string
      - in: query
        name: orderType
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: status
        type: integer
      - in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 活动订单列表
      tags:
      - LimitedActivityOrder
  /limitedActivity/order/refundOrder:
    post:
      parameters:
      - description: 订单ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 退款成功或已受理
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 活动订单退款(按未核销比例)
      tags:
      - LimitedActivityOrder
  /limitedActivity/order/verifyOrderByCodePublic:
    post:
      parameters:
      - description: 订单号
        in: query
        name: code
        required: true
        type: string
      responses:
        "200":
          description: 核销成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
                msg:
                  type: string
              type: object
      summary: 根据订单号核销活动订单(公开)
      tags:
      - LimitedActivityOrder
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...